+++ packages/flterm/lib/src/widgets/terminal_controller_impl.dart
@@ -112,6 +112,9 @@ class TerminalControllerImpl extends TerminalController
   }
 
   @override
+  set viewId(int? value) => _textInput.viewId = value;
+
+  @override
   TerminalConfig get config => _config;
 
   @override